move universal stream to singularity common
