+++ singularity_common/src/sap/mod.rs
@@ -1,5 +1,6 @@
 pub mod byte_stream;
 pub mod packet;
+pub mod universal_stream;
 
 // TODO: the two dependencies below should be removed
 // pub mod standard_features;
