+++ singularity_common/src/sap/packet.rs
@@ -36,103 +36,3 @@ pub fn join_id(id: IdType, inner_data: &[u8]) -> Vec<u8> {
 
     [id_bytes, inner_data].concat()
 }
-
-pub mod universal_client_stream {
-    use super::PacketTrait;
-    use crate::sap::byte_stream::ByteStream;
-
-    /// To be used by the client.
-    ///
-    /// NOTE: technically, I could have the generics be per-function,
-    /// but that might require more boilerplate for most cases
-    pub struct UniversalClientStream<Stream: ByteStream, Event: PacketTrait> {
-        stream: Stream,
-
-        event_queue: Vec<Event>,
-    }
-    impl<Stream: ByteStream, Event: PacketTrait> UniversalClientStream<Stream, Event> {
-        pub fn new(stream: Stream) -> Self {
-            Self {
-                stream,
-                event_queue: Vec::new(),
-            }
-        }
-
-        fn update_event_queue(&mut self) {
-            for raw_data in self.stream.try_iter_bytes() {
-                // currently disregard parsing errors (try_from_data errors),
-                // because it might just be an unsupported feature
-                if let Some(event) = Event::try_from_data(&raw_data) {
-                    self.event_queue.push(event);
-                };
-            }
-        }
-
-        /// Nonblocking
-        pub fn try_read_events(&mut self) -> Vec<Event> {
-            self.update_event_queue();
-
-            std::mem::take(&mut self.event_queue)
-        }
-
-        pub fn send_request<Request: PacketTrait>(&mut self, request: Request) {
-            self.stream.write_bytes(&request.to_data());
-        }
-    }
-}
-
-pub mod universal_server_stream {
-    use super::PacketTrait;
-    use crate::sap::byte_stream::ByteStream;
-    use std::marker::PhantomData;
-
-    /// To be used by the server.
-    ///
-    /// Represents connection to one client.
-    ///
-    /// REVIEW: naming, should I name this: `ClientHandler`?
-    ///
-    /// Right now, server and client socket are literally just the same with event and request switched.
-    /// I could abstract to just `UniversalStream<SendPacket, RecvPacket>`,
-    /// but I am preparing for queries and responses.
-    /// TODO: I could still abstract though.
-    pub struct UniversalServerStream<Stream: ByteStream, Event: PacketTrait, Request: PacketTrait> {
-        stream: Stream,
-
-        request_queue: Vec<Request>,
-
-        _r: PhantomData<Event>,
-    }
-    impl<Stream: ByteStream, Event: PacketTrait, Request: PacketTrait>
-        UniversalServerStream<Stream, Event, Request>
-    {
-        pub fn new(stream: Stream) -> Self {
-            Self {
-                stream,
-                request_queue: Vec::new(),
-                _r: PhantomData,
-            }
-        }
-
-        fn update_request_queue(&mut self) {
-            for raw_data in self.stream.try_iter_bytes() {
-                // currently disregard parsing errors (try_from_data errors),
-                // because it might just be an unsupported feature
-                if let Some(request) = Request::try_from_data(&raw_data) {
-                    self.request_queue.push(request);
-                };
-            }
-        }
-
-        /// Nonblocking
-        pub fn read_requests(&mut self) -> Vec<Request> {
-            self.update_request_queue();
-
-            std::mem::take(&mut self.request_queue)
-        }
-
-        pub fn send_event(&mut self, event: Event) {
-            self.stream.write_bytes(&event.to_data());
-        }
-    }
-}
